Mozik/requirements.txt
opencv-python
pillow
numpy
